requirements.txt
# Core dependencies
gradio>=4.0.0
requests>=2.31.0
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0

# Optional: Use faiss-gpu instead of faiss-cpu if you have CUDA
# faiss-gpu>=1.7.4

# Data processing
beautifulsoup4>=4.12.0
feedparser>=6.0.10
python-dateutil>=2.8.2

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0